Apply BigDecimal to dialogs

+++ app/src/main/java/com/example/traders/dialogs/buyDialog/BuyState.kt
@@ -1,15 +1,16 @@
 package com.example.traders.dialogs.buyDialog
 
 import com.example.traders.dialogs.DialogValidationMessage
+import java.math.BigDecimal
 
 data class BuyState(
     val amountToRound: Int = 2,
     val priceNumToRound: Int = 2,
-    val usdBalance: Double = 0.0,
-    val inputVal: Double = 0.0,
-    val minInputVal: Double = 10.0,
+    val usdBalance: BigDecimal = BigDecimal(0.0),
+    val inputVal: BigDecimal = BigDecimal(0.0),
+    val minInputVal: BigDecimal = BigDecimal(10.0),
     val isBtnEnabled: Boolean = false,
-    val cryptoToGet: Double = 0.0,
-    val usdLeft: Double = 0.0,
+    val cryptoToGet: BigDecimal = BigDecimal(0.0),
+    val usdLeft: BigDecimal = BigDecimal(0.0),
     val messageType: DialogValidationMessage = DialogValidationMessage.IS_EMPTY
 )
